llm-qa-module/requirements.txt
fastapi
uvicorn
requests
pydantic
langchain
langchain-huggingface
python-dotenv